datafeeder: move api response mappers to their own subpackage
